requirements.txt
pip
wheel
pyserial
PyQt5
pyqtgraph
numpy
